+++ ChangeLog.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file.
 ### Enhancements
 
 - Added RSA module (asymmetric public-key cryptography algorithm)
-    * Generate key pairs
+    * generate key pairs
     * store/load keys as PEM files
     * encrypt/decrypt
     * sign/verify